more stringent regex for desktop switching
A history entry looks like this:

firefox ***** Thu 8, 11:11

And the regex for desktop switching matches that string, and sends me to
desktop no. 8 instead of opening another firefox

A desktop entry looks like this

  8 firefox-bin  ----- urxvt (... and so on)

if it is currently inactive or

 *8 firefox-bin  ----- urxvt (... and so on)

if it is currently active. The new regex matches:

initially a space or a star
followed by a number,
followed by a single space,
followed by the rest of the line.